Fixed some things that annoyed me about earlier builds of this mod
On a side note, you should probably do something about the oredictionary
derpiness this mod has.

For instance, if I wanted to use a recipe that called for tomatoes for
any mod that registered an oredictionary definition for them, I could
use whole plants, which is not intended, and impossible for the most
part.